# Social-Distribtion-Engine-MelloxAI-SDE/specs/001-social-sde/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MelloxAI Social Distribution Engine (SDE)
  description: |
    Backend service for publishing and scheduling social media posts.

    ## Authentication
    All endpoints except `/healthz` require:
    - `Authorization: Bearer <SDE_API_TOKEN>`
    - `X-Signature: hex(hmac_sha256(SDE_SIGNING_SECRET, raw_body))`
    - `X-Timestamp`: Unix timestamp, reject if |now - ts| > 300s

    ## Error Response Format
    All errors follow this structure:
    ```json
    {
      "error_code": "ERROR_CODE",
      "detail": "Human-readable explanation",
      "how_to_fix": "Actionable next step",
      "field": "optional field name",
      "platform": "optional platform"
    }
    ```

    ## Key Concepts
    - **Post**: Logical publish request
    - **Target**: Individual platform publish attempt
    - **Idempotency**: Same idempotency_key returns existing post
  version: 1.0.0
  contact:
    name: MelloxAI Engineering
    url: https://mellox.ai
servers:
  - url: https://sde.mellox.ai/api/v1
    description: Production server
  - url: http://localhost:8000/api/v1
    description: Local development
paths:
...
  /jobs/{post_id}:
    get:
      summary: Get job status
      description: Returns full job details including all targets and event timeline
      operationId: getJob
      tags:
        - Jobs
      security:
        - BearerAuth: []
        - SignatureAuth: []
        - TimestampAuth: []
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Post UUID
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetailResponse"
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Cancel job
      description: |
        Cancel a pending job. Only allowed if no target is PUBLISHING or PUBLISHED.
        Returns 409 if job cannot be cancelled.
      operationId: cancelJob
      tags:
        - Jobs
      security:
        - BearerAuth: []
        - SignatureAuth: []
        - TimestampAuth: []
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Post UUID
      responses:
        "200":
          description: Job cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetailResponse"
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Job cannot be cancelled (already publishing or published)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs:
    get:
      summary: List jobs
      description: Paginated list of jobs with optional filters
      operationId: listJobs
      tags:
        - Jobs
      security:
        - BearerAuth: []
        - SignatureAuth: []
        - TimestampAuth: []
      parameters:
        - name: workspace_id
          in: query
          required: true
          schema:
            type: string
          description: Workspace identifier
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, PARTIAL, COMPLETED, FAILED, CANCELLED]
          description: Filter by status
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: Filter by created_at >= from
        - name: to
          in: query
          schema:
            type: string
            format: date-time
          description: Filter by created_at <= to
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 100
          description: Max results per page
        - name: cursor
          in: query
          schema:
            type: string
          description: Pagination cursor
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    SignatureAuth:
      type: apiKey
      in: header
      name: X-Signature
      description: HMAC-SHA256 signature of request body
    TimestampAuth:
      type: apiKey
      in: header
      name: X-Timestamp
      description: Unix timestamp (reject if > 5 min old)
  schemas:
...
    JobDetailResponse:
      type: object
      properties:
        post_id:
          type: string
          format: uuid
        workspace_id:
          type: string
        brand_id:
          type: string
        idempotency_key:
          type: string
        status:
          type: string
        scheduled_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        targets:
          type: array
          items:
            $ref: "#/components/schemas/JobTargetDetail"
        events:
          type: array
          items:
            $ref: "#/components/schemas/JobEvent"
    JobTargetDetail:
      type: object
      properties:
        target_id:
          type: string
          format: uuid
        platform:
          type: string
        status:
          type: string
        attempt_count:
          type: integer
        last_error_code:
          type: string
          nullable: true
        last_error_detail:
          type: string
          nullable: true
        platform_post_id:
          type: string
          nullable: true
        platform_post_url:
          type: string
          nullable: true
        published_at:
          type: string
          format: date-time
          nullable: true
    JobEvent:
      type: object
      properties:
        event:
          type: string
          enum:
            [post.queued, post.publishing, post.published, post.retrying, post.failed, post.partial]
        timestamp:
          type: string
          format: date-time
        target_id:
          type: string
          format: uuid
          nullable: true
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/JobSummary"
        next_cursor:
          type: string
          nullable: true
    JobSummary:
      type: object
      properties:
        post_id:
          type: string
          format: uuid
        status:
          type: string
        scheduled_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        target_count:
          type: integer
        published_count:
          type: integer
...
    ErrorResponse:
      type: object
      properties:
        error_code:
          type: string
        detail:
          type: string
        how_to_fix:
          type: string
...